requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
numpy==2.2.3
openpyxl==3.1.5
pandas==2.2.3
pyee==8.2.2
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
setuptools==75.8.0
six==1.17.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wheel==0.45.1
